# Conflicts:
#	crates/wysiwyg/src/composer_model/join_nodes.rs

+++ crates/wysiwyg/src/composer_model/replace_text.rs
@@ -168,7 +168,9 @@ where
 
             if range.is_empty() {
                 if !new_text.is_empty() {
-                    self.state.dom.append_child(DomNode::new_text(new_text));
+                    self.state
+                        .dom
+                        .append_at_end_of_document(DomNode::new_text(new_text));
                 }
                 start = 0;
             } else if let Some(starting_link_handle) = starting_link_handle {
@@ -222,7 +224,8 @@ where
         let action_list = self.replace_in_text_nodes(range.clone(), new_text);
 
         let (to_add, to_delete, _) = action_list.grouped();
-        let to_delete = to_delete.into_iter().map(|a| a.handle).collect();
+        let to_delete: Vec<DomHandle> =
+            to_delete.into_iter().map(|a| a.handle).collect();
 
         // We only add nodes in one special case: when the selection ends at
         // a BR tag. In that case, the only nodes that might be deleted are
@@ -239,7 +242,9 @@ where
         }
 
         // Delete the nodes marked for deletion
-        self.delete_nodes(to_delete);
+        if !to_delete.is_empty() {
+            self.delete_nodes(to_delete);
+        }
 
         // If our range covered multiple text-like nodes, join together
         // the two sides of the range.
